refactor(test): rename ListEntries gateway controller and tests
