+++ frontend/src/Infrastructure/Entries/ListEntriesGateway.ts
