+++ frontend/tests/Unit/Entries/BaseEntriesGatewayTest.ts
